refactor: controllers to follow RESTful conventions
